US97555 Remove image overlay
The `d2l-organization-info` component now supports showing additional information for future/past/inactive courses, so the image overlay is being retired. Unfortunately, there is still some remaining date-related code here, to allow us to set the `past-course` attribute on the card, as well as greying out inaccessible courses. There is probably a better way to do this, but the implementation of this behaviour in `d2l-organization-info` isn't quite complete yet anyway - we are still missing the status badge, which will slightly modify the available information.